Refactor DxfTablesSectionReader for modular table parsing
Refactored DxfTablesSectionReader to use generic, delegate-based methods for reading DXF tables and entries. Moved the class to the DxfStreamReader namespace, added dedicated entry readers for each table type, and improved error handling and notifications. Enhanced code clarity, maintainability, and extensibility.
